mark-vle-strands-agent/requirements.txt
strands-agents>=1.0.0
strands-agents-tools
boto3>=1.34.0
numpy>=1.24.0
python-dotenv>=1.0.0
gradio>=4.0.0
litellm>=1.0.0
flask>=3.0.0
flask-cors>=4.0.0
